Add implementation stubs and method to run tests
For readers who want to try implementing the Go
specifications for themselves, it's nice to start
with the stubs and a set of easy-to-run failing
tests.

This commit takes the implementation from v1,
changes all the implementation details into stubs.
It also takes the tests from v1 and v2, and makes
them easy to run with docker-compose.